pipeline/requirements.txt
anthropic>=0.40.0
feedparser>=6.0.11
python-dateutil>=2.9.0
python-dotenv>=1.0.0
requests>=2.32.0
